requirements.txt
flask==2.0.1
python-dotenv==0.19.1
requests==2.26.0
beautifulsoup4==4.10.0
transformers==4.18.0
torch==1.10.0
torchvision==0.11.1
torchaudio==0.10.0
matplotlib==3.4.3
numpy==1.23.5
lime==0.2.0.1